+++ repos/cache/src/commonMain/kotlin/dev/inmo/micro_utils/repos/cache/CacheRepo.kt
@@ -1,5 +1,11 @@
 package dev.inmo.micro_utils.repos.cache
 
-interface CacheRepo {
+interface InvalidatableRepo {
+    /**
+     * Invalidates its internal data. It __may__ lead to autoreload of data. In case when repo makes autoreload,
+     * it must do loading of data __before__ clear
+     */
     suspend fun invalidate()
 }
+
+typealias CacheRepo = InvalidatableRepo