+++ repos/cache/src/commonMain/kotlin/dev/inmo/micro_utils/repos/cache/KeyValueCacheRepo.kt
@@ -6,6 +6,7 @@ import dev.inmo.micro_utils.coroutines.withWriteLock
 import dev.inmo.micro_utils.pagination.*
 import dev.inmo.micro_utils.repos.*
 import dev.inmo.micro_utils.repos.cache.cache.KVCache
+import dev.inmo.micro_utils.repos.cache.util.actualizeAll
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.*
@@ -48,9 +49,7 @@ open class ReadKeyValueCacheRepo<Key,Value>(
         }
     }
 
-    override suspend fun invalidate() = locker.withWriteLock {
-        kvCache.clear()
-    }
+    override suspend fun invalidate() = kvCache.actualizeAll(parentRepo, locker = locker)
 }
 
 fun <Key, Value> ReadKeyValueRepo<Key, Value>.cached(
@@ -74,10 +73,6 @@ open class KeyValueCacheRepo<Key,Value>(
             kvCache.unset(it)
         }
     }.launchIn(scope)
-
-    override suspend fun invalidate() = locker.withWriteLock {
-        kvCache.clear()
-    }
 
     override suspend fun clear() {
         parentRepo.clear()